Consolidate Transformable as sealed trait
