+++ catalogue_pipeline/reindexer/src/main/scala/uk/ac/wellcome/platform/reindexer/services/CalmReindexTargetService.scala
@@ -6,7 +6,7 @@ import com.amazonaws.services.dynamodbv2.AmazonDynamoDB
 import com.gu.scanamo.Scanamo
 import com.twitter.inject.annotations.Flag
 import uk.ac.wellcome.metrics.MetricsSender
-import uk.ac.wellcome.models.CalmTransformable
+import uk.ac.wellcome.models.transformable.CalmTransformable
 import uk.ac.wellcome.reindexer.models.ScanamoQueryStream
 
 class CalmReindexTargetService @Inject()(
