+++ catalogue_pipeline/reindexer/src/test/scala/uk/ac/wellcome/platform/reindexer/locals/DynamoDBLocal.scala
@@ -4,7 +4,8 @@ import com.amazonaws.services.dynamodbv2.model._
 import com.gu.scanamo.Scanamo
 import org.scalatest.{BeforeAndAfterEach, Suite}
 import uk.ac.wellcome.models.transformable.miro.MiroTransformable
-import uk.ac.wellcome.models.{CalmTransformable, Reindex}
+import uk.ac.wellcome.models.Reindex
+import uk.ac.wellcome.models.transformable.CalmTransformable
 import uk.ac.wellcome.test.utils.DynamoDBLocalClients
 
 import scala.collection.JavaConversions._
