+++ catalogue_pipeline/transformer/src/main/scala/uk/ac/wellcome/transformer/parsers/CalmParser.scala
@@ -1,7 +1,7 @@
 package uk.ac.wellcome.transformer.parsers
 
 import com.twitter.inject.Logging
-import uk.ac.wellcome.models.CalmTransformable
+import uk.ac.wellcome.models.transformable.CalmTransformable
 import uk.ac.wellcome.utils.JsonUtil
 
 import scala.util.Try
