+++ catalogue_pipeline/transformer/src/main/scala/uk/ac/wellcome/transformer/transformers/MiroTransformableTransformer.scala
@@ -3,10 +3,8 @@ import java.io.InputStream
 
 import uk.ac.wellcome.models._
 import uk.ac.wellcome.models.transformable.Transformable
-import uk.ac.wellcome.models.transformable.miro.{
-  MiroTransformable,
-  MiroTransformableData
-}
+import uk.ac.wellcome.models.transformable.miro.MiroTransformable
+import uk.ac.wellcome.transformer.source.MiroTransformableData
 import uk.ac.wellcome.utils.JsonUtil
 
 import scala.io.Source
