+++ catalogue_pipeline/transformer/src/test/scala/uk/ac/wellcome/transformer/parsers/CalmParserTest.scala
@@ -1,7 +1,7 @@
 package uk.ac.wellcome.transformer.parsers
 
 import org.scalatest.{FunSpec, Matchers}
-import uk.ac.wellcome.models.CalmTransformable
+import uk.ac.wellcome.models.transformable.CalmTransformable
 import uk.ac.wellcome.transformer.utils.TransformableSQSMessageUtils
 
 class CalmParserTest
