--- common/src/main/scala/uk/ac/wellcome/models/transformable/miro/MiroTransformable.scala
@@ -1,17 +1,0 @@
-package uk.ac.wellcome.models.transformable.miro
-
-import uk.ac.wellcome.models.transformable._
-
-case class MiroTransformable(MiroID: String,
-                             MiroCollection: String,
-                             data: String,
-                             ReindexShard: String = "default",
-                             ReindexVersion: Int = 0)
-    extends Transformable
-    with Reindexable[String] {
-
-  val id: ItemIdentifier[String] = ItemIdentifier(
-    HashKey("MiroID", MiroID),
-    RangeKey("MiroCollection", MiroCollection)
-  )
-}
