+++ sierra_adapter/common/src/test/scala/uk/ac/wellcome/sierra_adapter/locals/DynamoDBLocal.scala
@@ -3,8 +3,8 @@ package uk.ac.wellcome.sierra_adapter.locals
 import com.amazonaws.services.dynamodbv2.model._
 import com.gu.scanamo.Scanamo
 import org.scalatest.{BeforeAndAfterEach, Suite}
-import uk.ac.wellcome.models.MergedSierraRecord
 import uk.ac.wellcome.dynamo._
+import uk.ac.wellcome.models.transformable.MergedSierraRecord
 import uk.ac.wellcome.test.utils.DynamoDBLocalClients
 
 import scala.collection.JavaConversions._
