+++ sierra_adapter/sierra_item_merger/src/main/scala/uk/ac/wellcome/platform/sierra_item_merger/services/SierraItemMergerWorkerService.scala
@@ -6,10 +6,10 @@ import grizzled.slf4j.Logging
 import io.circe.generic.extras.auto._
 import io.circe.parser._
 import uk.ac.wellcome.metrics.MetricsSender
-import uk.ac.wellcome.models.SierraItemRecord
 import uk.ac.wellcome.models.aws.SQSMessage
 import uk.ac.wellcome.sqs.{SQSReader, SQSReaderGracefulException, SQSWorker}
 import uk.ac.wellcome.circe._
+import uk.ac.wellcome.models.transformable.sierra.SierraItemRecord
 
 import scala.concurrent.ExecutionContext.Implicits.global
 import scala.concurrent.Future
