+++ sierra_adapter/sierra_items_to_dynamo/src/test/scala/uk/ac/wellcome/platform/sierra_items_to_dynamo/sink/SierraItemRecordMergerTest.scala
@@ -3,7 +3,7 @@ package uk.ac.wellcome.platform.sierra_items_to_dynamo.sink
 import io.circe.optics.JsonPath.root
 import io.circe.parser._
 import org.scalatest.{FunSpec, Matchers}
-import uk.ac.wellcome.models.SierraItemRecord
+import uk.ac.wellcome.models.transformable.sierra.SierraItemRecord
 import uk.ac.wellcome.utils.JsonUtil
 
 class SierraItemRecordMergerTest extends FunSpec with Matchers {
